requirements.txt
robotframework==6.0
robotframework-sapguilibrary==1.1
pyautogui==0.9.54
pyodbc==5.3.0
dotenv==0.9.9
rpaframework==31.1.0